firebase combined
