+++ src/front/js/pages/profile.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import Navbar from '../component/navbar';
 import { Footer } from "../component/footer";
 import { Link } from "react-router-dom";
-import { useAuth, upload } from "../../../../src/firebase.js";
+import { useAuth, upload } from  "../../firebase.js";
 import { BucketTest } from "../component/bucketTest";
